+++ src/components/day-page/NewDayButton.tsx
@@ -1,7 +1,7 @@
 import { useContext } from "react";
 import { Button } from "react-bootstrap";
-import { Action, newDayAction } from "../actions";
-import { MealDispatch } from "./MealDispatch";
+import { Action, newDayAction } from "../../actions";
+import { MealDispatch } from "../MealDispatch";
 
 export const NewDayButton = () => {
   const dispatch: React.Dispatch<Action> = useContext(MealDispatch);
